sentiment_analysis_backend/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
logging==0.4.9.6
MarkupSafe==2.1.5
numpy==2.1.0
praw==7.7.1
prawcore==2.4.0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
threadpoolctl==3.5.0
update-checker==0.18.0
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.4
